correção do timer, refatoração da tela Qual Desenho/Game
